// frontend/src/shared/types.ts
export interface Candle {
  time: number
  open: number
  high: number
  low: number
  close: number
}